+++ ocrspace/languages.py
