+++ ocrspace/test_pictures/hehehe differentiation.png
