+++ ocrspace/test_pictures/lmao imaginary numbers .png
