+++ ocrspace/test_pictures/secant and cosecant are losers.png
